+++ shopGunSdk/src/main/java/com/shopgun/android/sdk/ExternalClientIdStore.java
@@ -16,17 +16,17 @@
 
 package com.shopgun.android.sdk;
 
+import android.Manifest;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.os.Environment;
+import android.support.v4.content.ContextCompat;
 
 import com.shopgun.android.sdk.log.SgnLog;
 import com.shopgun.android.sdk.model.Session;
-import com.shopgun.android.sdk.utils.PermissionUtils;
 import com.shopgun.android.sdk.utils.Utils;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.RandomAccessFile;
 
 import static android.os.Environment.MEDIA_MOUNTED;
@@ -37,55 +37,28 @@ public class ExternalClientIdStore {
 
     private static final String CID_RANDOMJUNK = "randomjunkid";
 
-    public static void updateCid(Session s, Context c) {
-
-        String extCid = getCid(c);
-        String cid = s.getClientId();
-
-        // Previously used this hardcoded cid in beta, recover it
-        if (CID_RANDOMJUNK.equals(cid) || CID_RANDOMJUNK.equals(extCid)) {
+    public static void updateCid(Session s, ShopGun sgn) {
+        String extCid = getCid(sgn);
+        if (CID_RANDOMJUNK.equals(s.getClientId()) || CID_RANDOMJUNK.equals(extCid)) {
+            // Previously used this hardcoded cid in beta, recover it
             s.setClientId(Utils.createUUID());
-            saveCid(s.getClientId(), c);
-        } else if (cid == null) {
+        } else if (s.getClientId() == null) {
             // No ClientID is set, try to get from disk
             s.setClientId(extCid);
-        } else if (!cid.equals(extCid)) {
-            // ClientID have changed, write changes to disk
-            saveCid(cid, c);
         }
-
+        sgn.getSettings().setClientId(s.getClientId());
     }
 
-    private static void saveCid(String cid, Context c) {
+    public static String getCid(ShopGun sgn) {
 
-        File f = getCidFile(c);
-        if (f == null) {
-            return;
+        // First try SharedPrefs
+        String cid = sgn.getSettings().getClientId();
+        if (cid != null) {
+            return cid;
         }
 
-        FileOutputStream fos = null;
-        if (f.exists()) {
-            f.delete();
-        }
-        try {
-            fos = new FileOutputStream(f);
-            fos.write(cid.getBytes());
-            fos.flush();
-        } catch (IOException e) {
-            // Ignoring
-        } finally {
-            try {
-                fos.close();
-            } catch (Throwable t) {
-                // ignore
-            }
-        }
-
-    }
-
-    private static String getCid(Context c) {
-
-        File cidFile = getCidFile(c);
+        // Then try external storage
+        File cidFile = getCidFile(sgn.getContext());
         if (cidFile == null) {
             return null;
         }
@@ -110,12 +83,20 @@ public class ExternalClientIdStore {
             } catch (Exception e) {
                 // Ignore
             }
+
+            // Cleanup the cid file, we won't need it any more
+            deleteCidFile(sgn.getContext());
         }
 
         return null;
     }
 
-    private static boolean deleteCid(Context c) {
+    public static void clear(ShopGun sgn) {
+        deleteCidFile(sgn.getContext());
+        sgn.getSettings().setClientId(null);
+    }
+
+    private static boolean deleteCidFile(Context c) {
         File f = getCidFile(c);
         return f != null && f.exists() && f.delete();
     }
@@ -123,63 +104,23 @@ public class ExternalClientIdStore {
     private static File getCidFile(Context context) {
 
         if (MEDIA_MOUNTED.equals(Environment.getExternalStorageState()) &&
-                PermissionUtils.hasWriteExternalStorage(context)) {
+                ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                        == PackageManager.PERMISSION_GRANTED) {
 
-            File cacheDir = new File(Environment.getExternalStorageDirectory(), "cache");
-            if (!cacheDir.exists() && !cacheDir.mkdirs()) {
-                SgnLog.w(TAG, "External directory couldn't be created");
-                return null;
+            try {
+                File cacheDir = new File(Environment.getExternalStorageDirectory(), "cache");
+                if (!cacheDir.exists() && !cacheDir.mkdirs()) {
+                    SgnLog.w(TAG, "External directory couldn't be created");
+                    return null;
+                }
+                String fileName = context.getPackageName() + ".txt";
+                return new File(cacheDir, fileName);
+            } catch (Exception e) {
+                // If we are not allowed to access external storage
             }
-
-            String fileName = context.getPackageName() + ".txt";
-            return new File(cacheDir, fileName);
         }
 
         return null;
-    }
-
-    public static void test(Context c) {
-
-        long start = System.currentTimeMillis();
-
-        // Just clearing the prefs file
-        deleteCid(c);
-
-        boolean didFail = false;
-
-        String extCid = null;
-
-        // no CID has been obtained yet
-        Session s = new Session();
-        updateCid(s, c);
-        if (s.getClientId() != null || extCid != null) {
-            SgnLog.d(TAG, "ERROR: ClientId has been set: Session:" + s.getClientId() + ", mCid:" + extCid);
-            didFail = true;
-        }
-
-        String first = "fake_client_id";
-        s.setClientId(first);
-        updateCid(s, c);
-        if (!first.equals(s.getClientId())) {
-            SgnLog.d(TAG, "ERROR: ClientId changed: Session:" + s.getClientId() + ", mCid:" + extCid);
-            didFail = true;
-        }
-        if (!s.getClientId().equals(extCid)) {
-            SgnLog.d(TAG, "ERROR: ClientId mismstch: Session:" + s.getClientId() + ", mCid:" + extCid);
-            didFail = true;
-        }
-
-        String second = "new_fake_client_id";
-        s.setClientId(second);
-        updateCid(s, c);
-        if (!second.equals(s.getClientId()) || !s.getClientId().equals(extCid)) {
-            SgnLog.d(TAG, "ERROR: ClientId mismstch: Session:" + s.getClientId() + ", mCid:" + extCid);
-            didFail = true;
-        }
-
-        deleteCid(c);
-        SgnLog.d(TAG, "Test: " + (didFail ? "failed" : "succeded") + ", in " + (System.currentTimeMillis() - start));
-
     }
 
 }