+++ shopGunSdk/src/main/java/com/shopgun/android/sdk/pageflip/FragmentStatelessPagerAdapter.java
@@ -48,6 +48,7 @@ public abstract class FragmentStatelessPagerAdapter extends FragmentStatePagerAd
 
     /**
      * Return the Fragment associated with a specified position.
+     * @param position The position to create an item for
      */
     public abstract Fragment createItem(int position);